nghiahiep: add favourite service and controller

+++ src/main/java/org/learning/dlearning_backend/repository/FavoriteRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface FavoriteRepository extends JpaRepository<Favorite, Long> {
+public interface FavoriteRepository extends JpaRepository<Favorite, Integer> {
     Page<Favorite> findByUser(User user, Pageable pageable);
 
     List<Favorite> findByUser(User user);
